Cleaned up user refactor and refactored common dependencies
