+++ app/auth/dependencies.py
@@ -6,7 +6,7 @@ from fastapi.security import OAuth2PasswordBearer, SecurityScopes
 from jose import JWTError, jwt
 from pydantic import ValidationError
 
-from app.dependencies.common import get_user_by_username
+from app.dependencies import get_user_by_username
 from app.auth.models import TokenData
 
 SECRET_KEY = os.getenv("SECRET_KEY")
